+++ UVtools.Core/About.cs
@@ -7,7 +7,6 @@
  */
 
 using System;
-using System.IO;
 using System.Reflection;
 
 namespace UVtools.Core